fixed types and added support for many users on one account

+++ src/routes/accounts/handlers/delete-account.ts
@@ -12,23 +12,47 @@ export const DeleteAccountSchema = z.object({
 
 type Request = z.infer<typeof DeleteAccountSchema>
 
+/**
+ * This removes connection between the user and account
+ * and deletes the account if no other users are connected to it
+ */
 export const deleteAccount = async (req: ServerRequest<Request['body']>, res: ServerResponse) => {
   const { userId } = res.locals
   const { accountId } = req.body
 
-  // This removes connection between the user and account
-  // and deletes the account if no other users are connected to it
-  await prisma.accounts.delete({
+  await prisma.accounts.update({
     where: {
       id: accountId,
-
+    },
+    data: {
       users: {
-        some: {
+        disconnect: {
           id: userId,
         },
       },
     },
   })
+
+  // Get the account to check if there are any users left with that account
+  const account = await prisma.accounts.findMany({
+    include: {
+      users: true,
+    },
+
+    where: {
+      id: accountId,
+    },
+  })
+
+  const userCount = account[0]?.users.length
+
+  if (userCount === 0) {
+    await prisma.accounts.delete({
+      where: {
+        id: accountId,
+      },
+    })
+  }
 
   res.json({
     success: true,
